Move content to new metadata.py and storage.py
